pyproject.toml
[project]
name = "ros_tf_example"
authors = [{ name = "Daniel Duberg", email = "daniel@rerun.io" }]
requires-python = ">= 3.11"
version = "1.0.0"
readme = "README.md"
dependencies = [
    "rerun-sdk[datafusion]>=0.29.0-rc.4",
    "numpy",
    "pandas",
    "matplotlib",
    "pyarrow",
    "rosbags",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = [
    "linux-64", 
    "linux-aarch64",
    "osx-arm64", 
    "win-64",
]

[tool.pixi.pypi-dependencies]
ros_tf_example = { path = ".", editable = true }

[tool.pixi.tasks]
example = { cmd = "python3 ros_tf_example" }
tf_info = { cmd = "python3 ros_tf_example/tf_info.py" }

[tool.pixi.dependencies]
# quaternion = ">=2024.0.13,<2025"
# rrd_info = { cmd = "python3 ros_tf_example/rrd_info.py" }

[tool.ruff]
line-length = 150
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

ignore = [
    "E501", # Line too long.
    "F722", # Forward annotation false positive from jaxtyping. Should be caught by pyright.
    "F821", # Forward annotation false positive from jaxtyping. Should be caught by pyright.
]
